pyproject.toml
[tool.poetry]
name = "aws-transcribe-dockerized"
version = "0.0.1"
description = "A dockerizer implementation of aws transcribe"
authors = ["vizsatiz <vishnu@rootflo.ai>"]
license = "BSD 2.0 License"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
boto3 = "^1.35.14"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
